Bagpipe driver: use Neutron registry not an ML2 MD
This change makes the bagpipe driver use the Neutron
registry callbacks instead of a custom ML2 mech driver.

The now obsolete bgpvpn_notify mech driver still exists
to avoid breaking past configs, but it is empty and logs
a warning on startup.

With this change, the bagpipe driver also stops
preventing network deletion, because without a mech driver
we can't, which is good since it fixes bug 1489881.

This change also changes the behavior to ignore DHCP ports.

Implements: blueprint bagpipe-use-neutron-registry
Closes-Bug: 1489881
Partial-Bug: 1492349
Change-Id: I0dcfbadd222ad1f9768f4546118cb311fbfbdb51