OVS Agent extension for bagpipe driver
This change introduces an OVS L2 agent extension that will now be used
instead of the previous agent in ovs_bagpipe_neutron_agent.py which
was inheriting from the Neutron OVS agent in a crude way, requiring
duplicating the main() code.

We also remove mention of the old bagpipe-openvswitch agent
in the documentation and and adapt the devstack plugin so that
old local.conf will still work.

Change-Id: Ibf27e2b028efffa1e49207d39bfc9b13464180ad
Closes-Bug: 1492021
Closes-Bug: 1531459