Add tenant-id to subresources
When a tenant access a subresource, neutron will check that this
tenant is allowed to view this subresource, according to the policy.json.
The plugin must return the tenant_id parameter when a "show/list/delete"
action is performed, so that the policy framework can check authz.

This patch also set the network_id as a mandatory parameter
in the network_association table.

Change-Id: I59d2dbb5416b566dc7de3acb4fab3ed2ada8b78e
Fixes-Bug: #1512789