Client : adding the tenant-id if specified for a net association creation
Currently the client doesn't add the tenant_id in its request.
If the tenant_id is specified in the CLI, it has to be integrated
in the request sent by the client.

This is at least needed when a tenant wants to create a network
association on behalf of a tenant.

Closes-bug: #1512711

Change-Id: I9e1038372d1ba36ad6e3d6ca8339d4fa03471a53