Remove hardcoded utf8 coding for bgpvpns table
to align to the kilo code it is necessary to
let the database decide to use latin1 or utf8
as codec. Otherwise bgpvpn_network_associations
table cannot have foreign keys to networks
and bgpvpns. The reason is that for foreign keys
the column in the table and the column in the
foreign table has to have the same codec.

Change-Id: I2e732d8474f6769259e25d69b1b5269bebc9c26f
Closes-Bug: 1515987