bagpipe: port delete action on BEFORE_DELETE event
Details on the issue in bug #1517480.

We need to subscribe to port before_delete events, because
on after_delete events the port is already removed from the db
and we can't retrieve the binding:host_id information (which
is not passed in the event either).

This change also include a unit test against this bug, which also
happens to go through the ML2 RPC-to-registry codepaths so that
the effect of real registry events against our code are tested, thus
testing the validity of the expectation that the bagpipe driver code
has against ML2. This addresses bug #1493901.

One test is not passing, and thus commented out: see bug 1519258.

Closes-Bug: 1517480
Closes-Bug: 1493901

Change-Id: Icce89d98bcc4424529dcbe661fed3c1c8191418c