Add OpenContrail driver
That version uses the key/value store exposed by the OpenContrail
API but that's not efficient. The BGP/VPN resources and sub-association
need to be added to the OpenContrail data model.

That first PoC will be delivered to with the Liberty release but will be
replaced by a final one for Mitaka and next releases without any
migration script or method from the first PoC version.

Implements: blueprint opencontrail-bgpvpn-driver
Change-Id: I3795dd7b24afdac1cd5134c41b6b32ef8a888287