Devstack : configure tempest file during extra hook
The bgpvpn devstack plugin currently configures tempest during
the post-config stage. But the tempest.conf file is overwritten
by tempest during the extra stage of devstack, hence erasing
modifications made by the bgpvpn plugin.

To fix this issue, the modification of the tempest.conf file
must be done during the extra stage of devstack.

This patch also corrects the indentation of the devstack plugin.

Change-Id: Ia2e76f1c9aea93e481694b0821d1e5c3bed4da11
Closes-bug: #1548364