Make test jobs constrained
This change updates tox.ini and tools/tox_install.sh so that
we impose upper constraints on dependencies.

This is done by reusing what was done in neutron-lbaas.

(we also update our tools/tox_install to benefit from
a few improvements that have been brought to the one in
neutron-lbaas)

This change also declare our test dependency on pylint
in test-requirements.txt, rather than in tox.ini, so that
we don't endup with a pylint version that is inconsistent
with the one global-requirements would draw.

(cherry-picked from e8c379efd781244b6a95c52434a45f58bfd2f041)

Additionally, we stop having python-neutronclient stable/liberty
git branch declared as a test dependency, and now refer to
release 3.1.1 in requirements.txt (networking-bgpvpn actually
depends on a fix in this branch).

Change-Id: I904286860a5b8471f5da9312ee23bdf49004f1b0