Bring models in sync with migrations, add test
Bring the sqlalchemy models in sync with the alembic migrations.

Add the neutron standard model-migration sync test. The test is
added to the unit test env since bgpvpn does not have a
functional job yet. This requires that bgpvpn uses the
python-db-jobs template in the gate.

The neutron HasTenant model mixin includes an index for the
tenant_id column, but bgpvpn's alembic migrations never created
those indexes. So here we include a migration to add them.

(Also clean up some unused parameters and __init__ files.)

Closes-Bug: #1605608
Partial-Bug: #1604803

Depends-On: I3a8ea02b86846e9a8a7d7f2aee9700ca952b0248

Change-Id: I24d35f7dccab611143f4b0a7428ba7074f2adcc2