bagpipe: compatibility with Neutron routers
Add information on gateway MAC address to RPC notifications: when/if
a router interface is plugged provide the gateway MAC address
to the agent in RPCs.

This change includes the Neutron server-side part of the solution, the
agent-side evolution is covered in networking-bagpipe repo.

Additionally, we remove a Router/Network check that has become
redundant now that the check has been introduced in the plugin
( Ie1f6a733e8a16b8b41e89c5a82637d7cb03bd423 ).

Partial-Bug: 1627645
Change-Id: I89c07d9166f23d910542a825c5000fdd22580217