undo some of the incorrect changes for prepping Newton
Most of the changes in d08892589133fe08e1aac88f0b478381ecd332e4 should
have been made on the branch after it was created. The update to
requirements was the only one that should have been made in master
before the branch.

The change to .gitreview and tox.ini will be made automatically by the
branch script. The other changes should be done by hand in
stable/newton.

Change-Id: Ief5c73ab4d1cd5fd032dede21f6a244bfa5777c2
Signed-off-by: Doug Hellmann <doug@doughellmann.com>