Temporary workaround to our gate issues
Our tempest test scheduled on osic-cloud1 all get stuck
in requirements installation.

The problem was diagnosed (fungi, tmorin) to be related
to the fact that the neutron devstack code for setting
up routers is installing via br-ex a route to SUBNETPOOL_PREFIX_V4,
which defaults to 10.0.0.0/16, which in the case of
osic-cloud1 contains the gateway IP to reach the Internet
(10.0.0.1):

fungi@ubuntu-xenial-osic-cloud1-s3500-5253075:~$ ip ro sh
default via 10.0.0.1 dev ens3 onlink
10.0.0.0/16 via 172.24.5.11 dev br-ex
10.0.0.0/8 dev ens3  proto kernel  scope link  src 10.5.131.187
172.24.5.0/24 dev br-ex  proto kernel  scope link  src 172.24.5.1

This change is a temporary fix for networking-bgpvpn that
sets SUBNETPOOL_PREFIX_V4 to a prefix *not* containing 10.0.0.1.

Change-Id: I5e4b99c6eb06837be7e1b5a9ed456f57ab6f9514