Cap dependency on networking-bagpipe + reno fix
Or else we draw Ocata's networking-bagpipe, which is not compatible
with Newton's neutron (eg. [1]).

Issue detected thanks to periodic jobs.

This change has also the content of reno fix Ida99d7282cdf25cb287d37c484d90e1e19651747
squashed in: leave branch unspecified for "current" series.
The branch needs to default to whatever is checked out for this file to
trigger testing the build of unmerged release notes. Otherwise it
literally only looks at things that have merged into the master branch,
which allows a release note file with bad markup to merge and break the
build.

[1] http://logs.openstack.org/periodic-stable/periodic-networking-bgpvpn-python27-newton/d91cbf9/console.html#_2017-03-01_06_12_03_108318

Co-authored-by: Doug Hellmann <doug@doughellmann.com>
Change-Id: Ic109c11cbc72da4f206606995be156fc0f664266