leave branch unspecified for "current" series
The branch needs to default to whatever is checked out for this file to
trigger testing the build of unmerged release notes. Otherwise it
literally only looks at things that have merged into the master branch,
which allows a release note file with bad markup to merge and break the
build.

Change-Id: I22b0386e1ad50e0e68ab590bc822be09200ed557
Signed-off-by: Doug Hellmann <doug@doughellmann.com>