leave branch unspecified for "current" series
The branch needs to default to whatever is checked out for this file to
trigger testing the build of unmerged release notes. Otherwise it
literally only looks at things that have merged into the master branch,
which allows a release note file with bad markup to merge and break the
build.

Change-Id: I95f85a94fc3fe899764503fbb78e5cfa18c6d2e3
Signed-off-by: Doug Hellmann <doug@doughellmann.com>