Switch DB and driver precommit methods for delete and update
Deleting network and router assoc calls need to execute precommit and
postcommit checks. The method deleted the object in the DB before making
precommit checks.

With this change it makes precommit checks before deleting the object, which
allows the drivers to read the DB with the object being present at delete
precomit time.

The same change is applied to updates.

Co-Authored-By: Omar Sanhaji <sanhaji.omar@gmail.com>
Co-Authored-By: Thomas Morin <thomas.morin@orange.com>

Closes-Bug: #1709647
Partial Implements: blueprint bagpipe-bgpvpn-ovo

Change-Id: I4a191c1c6a1c5281cd18e07dcee534dc84338acc