dashboard: fix bug about route target validation
- When a route target is incorrect in the create or update form
a popup message should indicate that validation of this field failed.
This message doesn't display and instead a generic error is returned.
It's a mistake to convert the route_target which is charfield type to
a list in the clean method.
This, need to be done after in the handle method when the datas are "cleaned".

- Fix also bad behavior when a admin user update a BGPVPN. In some cases
update doesn't work when the RT field are edited.

Closes-Bug: 1716941
Closes-Bug: 1716689
Change-Id: I5fa5a15ddf90bc2a38645c54132cc6b57e9b797a