routes-control: add advertise_extra_routes to router_association
This change adds the DB, service plugin and driver hooks
to implement the new advertise_extra_routes attribute of BGPVPN
router associations.

It also adds the required methods in DB, service plugin
and driver hooks, to allow updating a router_association
resource (which was not previously proposed in the base BGPVPN
API extension).

Change-Id: I58014ea8629c1c613d1bbab1285d6ff5e25aaf9a
Partially-Implements: blueprint routes-control