Filter duplicated RTs in compiled list
The code does not filter duplicates: e.g. an RT that is both
in route_targets and import_targets will be present twice in the
RPCs between neutron server and agents, and agents will also pass
this RT twice to bagpipe-bgp.

This patch fix avoid duplicated RTs in compiled Import Targets and
Export Targets lists.

Change-Id: Ia5b76235c75f005e3c53f7cddc8e085376f5ebe1
Closes-Bug: 1719589