dashboard: impossible to add a bgpvpn with a empty route target
If the route targets, import targets or export targets are empty in the
django form, the creation or modification of the bgpvpn failed.
The neutron server return a error (Invalid input for export_targets.
Reason: '' is not a valid input.)
The route_targets are not correctly formatted and a empty string is sent
instead an empty list when the fields route_targets are empty in the form.

Change-Id: I3b233f68c12795713daf6c1a00042730eea84c56
Closes-Bug: 1730678