make get_extended_resources class methods
Currently the bgpvpn UTs are failing with neutron-lib/master [1] due to
[2] which is in neutron-lib master, but not yet released. While removing
the get_extended_resources in bgpvpn could fix [1], it causes problems
with sub-resource extensions in [3]. Therefore in order to fix [1], but
also not break [3] we need a temporary fix until [2] is released in the
next version of neutron-lib at which point the get_extended_resources
can be removed herein and [3] will still work.

This patch proposes we make the get_extended_resources classmethods as
a temporary mitigation. Once the next version of neutron-lib is released
and we consume it here, we can remove the get_extended_resources all
together and things will just work (based on my testing).

[1] http://logs.openstack.org/periodic/git.openstack.org/openstack/
networking-bgpvpn/master/legacy-periodic-py35-with-neutron-lib-master/
c356255/job-output.txt.gz
[2] I8ae11633962a48de6e8559b85447b8c8c753d705
[3] I58014ea8629c1c613d1bbab1285d6ff5e25aaf9a

Change-Id: Ifbaec5debf3e9ab024f3f877e678c1679c0f9e50