bagpipe v2 driver
The behavior of the current "v1" bagpipe driver is to
produce both the old and new RPCs to not prevent rolling
upgrades.

This change introduces a bagpipe v2 driver that does not
produce old RPCs and only relies on OVO push/pull RPCs.

To avoid code duplication of the code for new RPCs,
the current driver is refactored to inherits from the
v2 driver the code for the new OVO push/pull RPCs.

The current v1 driver is also modified to raise a warning
indicating its future removal.

This change also extends the previous unit test
coverage for the production of push-based RPCs.

Change-Id: I708ec1e4d8a8c27ffe8c4be172812e0a9aed54b4