[dashboard] Modify bgpvpn network associations
This patch adds a button "create network association" in bgpvpn panels. It
opens a form to add only one network association.

The bgpvpn overview details is more complete. There is now a
network associations tab which list all the network associations. It's
possible to delete each network association. It's inspired by the network
panel which include subnets and ports tabs on the overview of a network.

It will be easier in the future to add others associations as routers and
ports with the same logical with new tab in bgpvpn overview details.

Change-Id: I6280b53558b6b52081ee6d629e8f3963d8dfc346