tempest: use IP address ranges based on environment
Some of the scenario tests seems to fail due to the fact
that we use IP addresses that may be reachable in some CI cloud
providers. When/if this happens, our negative connectivity checks
fail to give the expected results: trying to reach a BGPVPN IP that
should not be reachable anymore, succeeds because this IP is reachable
via Neutron's router in the cloud provider address space.

To avoid this issue, when the SUBNETPOOL_PREFIX_V4 environement
variable is set, which is the case in a devstack environment, then
our scenario job will use /24s in this space.

Closes-Bug: 1749388
Change-Id: I79fd8b4e0f8643170e6e9b2474845114ceece99a