Avoid tox-install.sh
Instead of using tox-install.sh, we can install horizon and neutron
etc. now directly from pypi - or use the current branch in CI system using
tox-siblings.

Create doc/requirements.txt for doc requirements as well.

Update tox.ini for this change and add projects to test-requirements
file.

Note this is one step in fixing publishing of docs for this repository.

Depends-On: https://review.openstack.org/552865
Depends-On: https://review.openstack.org/554297
Change-Id: I13e61fffca48b6c004e028f0da8a1d84a2337e10