[dashboard] Modify bgpvpn router associations
This patch add a button "create router association" in bgpvpn panels.
It open a form to add only one network association with optionals parameters.

The bgpvpn overview details is more complete. There is
now a router associations tab which list all the router
associations. It's possible to delete each router
association and to update it.

Change-Id: Idd5af2ce2d4d864882c6191d648d16d9d9bcf35f
Partially-Implements: blueprint routes-control