tempest: dynamic RT allocation
The BGP RTs used in the tests were initially constants, reused
by the different tests.  This is safe only if tests are not
run in parallel, but not when they are concurrently run, because
some routes advertised by one test may be imported in routing
tables from another test.

To solve this, we need to dynamic allocation of RTs so that
different tests use different RTs, to ensure isolation
between the different tests.

Change-Id: Ib1b461a7f6d6d949993c07129f24fbffe92990bb