doc: update python API client documentation
The documentation was still refering to the
old methods provided by the networking-bgpvpn
dynamic extension to neutronclient instead of
the newer and more complete methods that have
been in neutronclient since Ocata release.

This change updates the documentation and examples,
and adds code to document the method for port
associations.

Partial-Bug: 1761691
Change-Id: Idd572530d64cd65cc0a606515bcf7f056c2b8745