dashboard: Fix test failures caused by django test runner
Recently horizon test configuration was changed to use
the Django default test runner because nose is in maintenance mode
for the past several years [1]. Django testcase and testtools are
incompatible, so bgpvpn_dashboard testcases now fail.

In addition, bgpvpn_dashboard/test/test.py is unnecessary
and it causes the test failure too, so this file will be
dropped by this commit.

[1] horizon commit 1f80d94459856a8c477310cc0fe4b0e165d8c0c1

Change-Id: I45ff876871d7dcb82f284061b3a7225dd65c771b
Closes-Bug: #1776223