switch to stestr
Update to new PTI:
https://governance.openstack.org/tc/reference/pti/python.html#python-test-running

This change also wraps in a change to requirements.txt that
point pre-rocky neutron pypi release, which is the only
way I found to have the functional job get a fresh-enough
neutron (it's not a tox job and hence does not benefit
from the tox-siblings magic). This also reflects better what
is the real dependency we have (when rocky will ship we will
have neutron>=13.0.0.0).

Change-Id: I868cf65aa5ff0e55d0d45f36808c76286e4e7fd3