Drop nose dependencies
Horizon test setting no longer depends on nose and related stuffs.
This commit cleans up nose related dependencies.

The change was made in horizon Rocky-3,
so horizon 14.0.0.0b3 is required at least.

Filenames of test modules need to start with test_*
so that Django test runner detects them.
There is no test defined but it would be nice to adjust the filename.

Change-Id: I794e10b2c3c2c7004f7b58ff8c8904d4537a23d9