use payloads for ROUTER_INTERFACE BEFORE_DELETE events
This patch switches over to the payload style of callbacks by updating
neutron to callbacks to handle a DBEventPayload for BEFORE_DELETE events
of ROUTER_INTERFACE resources. See the depends on patch for more
details.

This change also disables legacy-tempest-dsvm-networking-bgpvpn-bagpipe
that is currently broken because of bug 1807152 - since both this change
and that bug break the gate, we have to squash.

Change-Id: Ieda7d74aec8ccad4b22e19b0470237fa70885714
Depends-On: https://review.openstack.org/#/c/598347/