Add required projects
Stable periodic jobs started to fail [1] [2] with errors that they are
missing required projects as soon as old job settings were removed [3]
from project-config repository. The problem is only present in queens
and pike, on the other stable branches the configurations are OK.

So networking-bagpipe, horizon and neutron added as they were present
in the old settings.

[1] http://logs.openstack.org/periodic-stable/git.openstack.org/openstack/networking-bgpvpn/stable/pike/build-openstack-sphinx-docs/594eeba/job-output.txt.gz#_2018-10-13_06_33_35_696317
[2] http://logs.openstack.org/periodic-stable/git.openstack.org/openstack/networking-bgpvpn/stable/queens/openstack-tox-py27/4af2104/job-output.txt.gz#_2018-10-13_06_24_06_925732
[3] https://review.openstack.org/#/c/597865/

Change-Id: Ia9f5fa177771ead870c13546aa870ee64b3445be