use rpc from neutron-lib
We recently removed neutron.common.rpc [1] as the rpc module lives in
neutron-lib now. This project was missed because I didn't find
neutron.common.rpc as an import.

This patch switches rpc uses over to neutron-lib.

This patch also wraps in a needed update to lower-constraints.txt
(cannot merge separately, because both changes are needed to
unbreak the CI)

[1] https://review.openstack.org/#/c/634790/

Co-Authored-By: Akihiro Motoki <amotoki@gmail.com>
Change-Id: I42fe0a9a05ead681fd2a579650218c6140e3bb3a