Add local bindep.txt
As it was announced [1] global bindep-fallback.txt was removed and now
projects need to have a local bindep.txt to be able to install binary
dependencies for testing.

In test jobs the script tools/test-setup.sh is called which requires
mysql and postgres servers and clients to be installed.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2019-June/007272.html

Change-Id: I6b7afd3f6d0b45337f87e13352ab1095d51a23a3
(cherry picked from commit 4655630adc4ce3b6c69d3deffb397c36d67fab04)