[ussuri][goal] Drop python 2.7 support and testing
OpenStack is dropping the py2.7 support in ussuri cycle.

networking-bgpvpn is ready with python 3 and ok to drop the
python 2.7 support.

Also this patch update minimum tox version and remove 'pypy' entry
as it is no longer run in tox.ini file.
Complete discussion & schedule can be found in
- http://lists.openstack.org/pipermail/openstack-discuss/2019-October/010142.html
- https://etherpad.openstack.org/p/drop-python2-support

Ussuri Communtiy-wide goal:
https://governance.openstack.org/tc/goals/selected/ussuri/drop-py27.html

Change-Id: Id8046ef85da72207097fef52b7966a4b3a067704