stable only: Fix failing jobs and make l-c non-voting
Fix pep8 with missing isort, networking-bgpvpn-dsvm-functional with
ensuring tox, and make lower-constraints job non-voting.
Remove tools/test-setup.sh and let zuul setup database.

Default nodeset was set to focal in zuul and requirements-check job
started to fail in stable/stein. This patch sets ubuntu-bionic as
nodeset for the job to make it pass. doc/requirements.txt has to be
updated with the correct sphinx requirement constraints to make the
requirements-check job pass.

Change-Id: I3d49c6dcf004d1ae216d20aa9b59e199c21d61f1